backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pandas>=1.5.0
numpy>=1.21.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0
